Add explicit instructions for where to get images
In lesson 1, I'm not sure it's clear where to get the image files from.
This commit gives the student instuctions for downloading the images and
adding them to an /images folder.  It also makes sure the images
attributes are rendered in a list (before they weren't as the markdown
was missing a line before the list).