Clearer title for section about the nested element selector.
The previous headings in the selector section all refer to the new
type of selector they introduce (element, class, etc) except for this
one which (I think) refers to the effect the css ‘display’ property has
on positioning elements.

The proper name for this type of selector looks to be the ‘descendant’
selector[1], but I think ‘nested’ is a better word for us here as it
is a term that has already been introduced in lesson 1[2] and is used
later on in this paragraph.

[1] http://www.w3.org/TR/CSS2/selector.html#pattern-matching
[2] https://github.com/codebar/tutorials/blob/24e6c063909296e2faa3e147cc5b3aab83f2b332/html/lesson1/tutorial.md#nesting-elements