JS Lesson 4 update
The GitHub API changed and removed Gravatar URLs completely:
https://developer.github.com/changes/2014-09-05-removing-gravatar-id/

My students also had some trouble following the instructions given the very
loose structure. I added a bit more explanation around JSON and guidance for
implementing the XHR call method.

A student also mentioned that is what hard to follow because they didn't know
what the end goal of the tutorial is, so I added a link to the finished example
to the top.