Rework js tutorial #1
This is a substantial rework of the text, and a few things have moved
between tutorials #1 and #2. Major changes:

 - if statements, booleans, and operators other than + have moved to #2
 - objects have moved to #1
 - more detailed explanations, slower pacing early on
 - much deeper focus on functions

The things which were moved around are in aid of a larger refactor:
after this change, students should be able to manage js tutorials #3
through #6 using only the material learned in #1. This lets them jump
into building working websites faster.

Things which were removed from tutorial #1 have been added to #2 for
now just to have them somewhere, but a near-term goal should be to
break #2 into smaller chunks and introduce them as part of other
exercises, where students can be building things instead of reading
about them.