Change exercise 2 gist download link to new version
Change html code in exercise 2 to make `input` element self closing. This is because unnecessary closing tag was causing a bug in the jQuery.

I have changed the url for the gist as I was unable to edit the gist directly. Please feel free to edit despo's gist yourself instead if you prefer.