Changed to Python3 + added lists & dicts
This started as an update from v2 to v3 and balooned into the addition
of sections on types & casting, lists & dictionaries, and the breaking
out of the installation section as a separate tutorial.

Additionally, I changed some of the language and swapped out the use of
`range()` for `len()` since it made less sense to use it at a stage
where the user hadn't yet been introduced to lists.

Changed the title of lesson1 to 'Installation Guide'