Clarify HTML entity example
We replaced the double quotation marks example with an ampersand example, using the opening line to Edward Lear's poem 'The Owl and the Pussy-cat'.  The double quotation marks do not strictly speaking require an HTML entity in the previous example whereas an ampersand always requires an HTML entity.

We also removed the 'em' because we already have 'strong' as a follow on from a prior point and adding 'em' does not add anything to the learning of this particular example.