Use `>>> ` and `print()` uniformly
One of my students last night got confused with the tutorial where we
weren't directly calling `print()` and yet stuff was appearing on the
screen.  She wasn't sure if simply typing out the variable was somehow
*doing* something.  Using `print()` explicitly everywhere should clear
this up.

Similarly, `>>> ` was only being used in parts of the document, leaving
it unclear as to whether this is something that needs to be typed, or if
perhaps the absence of it meant they weren't in the REPL.

This commit addressed both issues.