Added explanation on line 125, Ruby tutorial 4
I had trouble understanding the expected outcome when changing the method from protected to private, as the wording implied a different outcome.  I got stuck working through this part of the tutorial, and after discussing with my coach, I've changed the wording to be clearer for other students, stating which error is expected and why.